requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
starlette==0.27.0

# Machine Learning Core
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
joblib==1.3.2
scipy==1.15.3

# NLP and Transformers
transformers==4.56.1
sentence-transformers==5.1.0
huggingface-hub==0.34.4
tokenizers==0.22.0
safetensors==0.6.2
sentencepiece==0.2.1
nltk==3.9.1
regex==2025.9.1

# Database and Analytics
duckdb==0.9.2
sqlalchemy==2.0.23
alembic==1.13.1
datasets==2.15.0
pyarrow==21.0.0
pyarrow-hotfix==0.7

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==45.0.7
bcrypt==4.3.0

# HTTP and Async
httpx==0.25.2
aiofiles==23.2.1
websockets==12.0
aiohttp==3.12.15
requests==2.32.5

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Data Processing and Utilities
openpyxl==3.1.2
python-dateutil==2.8.2
python-dotenv==1.1.1
tqdm==4.67.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# System Monitoring
psutil==5.9.6

# Code Quality Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==6.0.1
ruff==0.12.12
pylint==3.3.8

# Production Deployment
gunicorn==21.2.0

# System Dependencies (auto-installed by above packages)
# These are typically installed automatically but listed for completeness
certifi==2025.8.3
urllib3==2.5.0
charset-normalizer==3.4.3
idna==3.10
click==8.2.1
packaging==25.0
filelock==3.19.1
fsspec==2023.10.0
pyyaml==6.0.2
jinja2==3.1.6
markupsafe==3.0.2
typing-extensions==4.15.0
annotated-types==0.7.0
pydantic-core==2.14.1
anyio==3.7.1
sniffio==1.3.1
h11==0.16.0
watchfiles==1.1.0
uvloop==0.21.0
